# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gOpenID Admin API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Access and refresh token pair
  /api/auth/refresh:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Rotated token pair
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...